Refactor: Extract dynamic methods to a fluent class
